requirements.txt
APScheduler==3.6.3
asgiref==3.8.1
cachetools==4.2.2
certifi==2024.7.4
Django==5.0.7
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
pytest-django==4.8.0
python-decouple==3.8
python-telegram-bot==13.7
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
tornado==6.4.1
typing_extensions==4.12.2
tzlocal==2.1
